requirements.txt
# Core ML/NLP
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
sentence-transformers>=2.3.0
tokenizers>=0.15.0

# Retrieval
faiss-cpu>=1.7.4

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.1

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Experiment tracking
wandb>=0.16.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.1
hydra-core>=1.3.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
